@@ -1,4 +1,4 @@
-var parse = require('./parse');
+import parse from './parse.js';
 
 /**
  * Converts a string to command line args, in particular
@@ -38,7 +38,7 @@ function stringToArgs(string) {
   return args;
 }
 
-module.exports = {
+export default {
   parse: function (argv) {
     if (typeof argv === 'string') {
       argv = stringToArgs(argv);
