@@ -15,11 +15,11 @@ nodemon can be run in a number of ways:
 
 */
 
-var fs = require('fs');
-var path = require('path');
-var existsSync = fs.existsSync || path.existsSync;
+import fs from 'fs';
+import path from 'path';
+const existsSync = fs.existsSync || path.existsSync;
 
-module.exports = parse;
+export default parse;
 
 /**
  * Parses the command line arguments `process.argv` and returns the
