@@ -1,4 +1,6 @@
-var ignoreRoot = require('ignore-by-default').directories();
+import ignoreByDefault from 'ignore-by-default';
+
+var ignoreRoot = ignoreByDefault.directories();
 
 // default options for config.options
 const defaults = {
@@ -31,4 +33,4 @@ if (/--(loader|import)\b/.test(nodeOptions)) {
   delete defaults.execMap.ts;
 }
 
-module.exports = defaults;
+export default defaults;
