@@ -1,10 +1,13 @@
-const path = require('path');
-const fs = require('fs');
-const existsSync = fs.existsSync;
-const utils = require('../utils');
+import path from 'path';
+import fs from 'fs';
+import utils from '../utils/index.js';
+import { createRequire } from 'module';
 
-module.exports = exec;
-module.exports.expandScript = expandScript;
+const require = createRequire(import.meta.url);
+const existsSync = fs.existsSync;
+
+export default exec;
+export { expandScript };
 
 /**
  * Reads the cwd/package.json file and looks to see if it can load a script
