@@ -1,15 +1,19 @@
-var debug = require('debug')('nodemon');
-var fs = require('fs');
-var path = require('path');
-var exists = fs.exists || path.exists;
-var utils = require('../utils');
-var rules = require('../rules');
-var parse = require('../rules/parse');
-var exec = require('./exec');
-var defaults = require('./defaults');
+import Debug from 'debug';
+import fs from 'fs';
+import path from 'path';
+import utils from '../utils/index.js';
+import rules from '../rules/index.js';
+import parse from '../rules/parse.js';
+import exec from './exec.js';
+import defaults from './defaults.js';
+import { createRequire } from 'module';
 
-module.exports = load;
-module.exports.mutateExecOptions = mutateExecOptions;
+const debug = Debug('nodemon');
+const require = createRequire(import.meta.url);
+const exists = fs.exists || path.exists;
+
+export default load;
+export { mutateExecOptions };
 
 var existsSync = fs.existsSync || path.existsSync;
 
