@@ -1,12 +1,13 @@
-var fs = require('fs');
-var path = require('path');
-const supportsColor = require('supports-color');
+import fs from 'fs';
+import path from 'path';
+import { fileURLToPath } from 'url';
+import supportsColor from 'supports-color';
 
-module.exports = help;
-
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 const highlight = supportsColor.stdout ? '\x1B\[$1m' : '';
 
-function help(item) {
+export default function help(item) {
   if (!item) {
     item = 'help';
   } else if (item === true) { // if used with -h or --help and no args
