@@ -1,4 +1,7 @@
-module.exports = {
-  run: require('./run'),
-  watch: require('./watch').watch,
+import run from './run.js';
+import { watch } from './watch.js';
+
+export default {
+  run,
+  watch,
 };
