@@ -1,11 +1,13 @@
-const minimatch = require('minimatch');
-const path = require('path');
-const fs = require('fs');
-const debug = require('debug')('nodemon:match');
-const utils = require('../utils');
+import minimatch from 'minimatch';
+import path from 'path';
+import fs from 'fs';
+import Debug from 'debug';
+import utils from '../utils/index.js';
 
-module.exports = match;
-module.exports.rulesToMonitor = rulesToMonitor;
+const debug = Debug('nodemon:match');
+
+export default match;
+export { rulesToMonitor };
 
 function rulesToMonitor(watch, ignore, config) {
   var monitor = [];
