@@ -1,23 +1,23 @@
-var debug = require('debug')('nodemon:run');
-const statSync = require('fs').statSync;
-var utils = require('../utils');
-var bus = utils.bus;
-var childProcess = require('child_process');
-var spawn = childProcess.spawn;
-var exec = childProcess.exec;
-var execSync = childProcess.execSync;
-var fork = childProcess.fork;
-var watch = require('./watch').watch;
-var config = require('../config');
+import Debug from 'debug';
+import { statSync } from 'fs';
+import utils from '../utils/index.js';
+import { spawn, exec, execSync, fork } from 'child_process';
+import { watch } from './watch.js';
+import config from '../config/index.js';
+import psTree from 'pstree.remy';
+import path from 'path';
+import signals from './signals.js';
+import undefsafe from 'undefsafe';
+import { release } from 'os';
+
+const debug = Debug('nodemon:run');
+const bus = utils.bus;
+const osRelease = parseInt(release().split('.')[0], 10);
+
 var child = null; // the actual child process we spawn
 var killedAfterChange = false;
 var noop = () => {};
 var restart = null;
-var psTree = require('pstree.remy');
-var path = require('path');
-var signals = require('./signals');
-const undefsafe = require('undefsafe');
-const osRelease = parseInt(require('os').release().split('.')[0], 10);
 
 function run(options) {
   var cmd = config.command.raw;
@@ -107,10 +107,6 @@ function run(options) {
     inBinPath = statSync(`${binPath}/${executable}`).isFile();
   } catch (e) {}
 
-  // hasStdio allows us to correctly handle stdin piping
-  // see: https://git.io/vNtX3
-  const hasStdio = utils.satisfies('>= 6.4.0 || < 5');
-
   // forking helps with sub-process handling and tends to clean up better
   // than spawning, but it should only be used under specific conditions
   const shouldFork =
@@ -118,8 +114,7 @@ function run(options) {
     !inBinPath &&
     !(firstArg.indexOf('-') === 0) && // don't fork if there's a node exec arg
     firstArg !== 'inspect' && // don't fork it's `inspect` debugger
-    executable === 'node' && // only fork if node
-    utils.version.major > 4; // only fork if node version > 4
+    executable === 'node'; // only fork if node
 
   if (shouldFork) {
     // this assumes the first argument is the script and slices it out, since
@@ -130,7 +125,7 @@ function run(options) {
     const forkOptions = {
       env: env,
       stdio: stdio,
-      silent: !hasStdio,
+      silent: false,
     };
     if (utils.isWindows) {
       forkOptions.windowsHide = true;
@@ -288,26 +283,11 @@ function run(options) {
 
     // swallow the stdin error if it happens
     // ref: https://github.com/remy/nodemon/issues/1195
-    if (hasStdio) {
-      child.stdin.on('error', () => {});
-      process.stdin.pipe(child.stdin);
-    } else {
-      if (child.stdout) {
-        child.stdout.pipe(process.stdout);
-      } else {
-        utils.log.error(
-          'running an unsupported version of node ' + process.version
-        );
-        utils.log.error(
-          'nodemon may not work as expected - ' +
-            'please consider upgrading to LTS'
-        );
-      }
-    }
+    child.stdin.on('error', () => {});
+    process.stdin.pipe(child.stdin);
 
     bus.once('exit', function () {
       if (child && process.stdin.unpipe) {
-        // node > 0.8
         process.stdin.unpipe(child.stdin);
       }
     });
@@ -559,4 +539,4 @@ if (!utils.isWindows) {
   });
 }
 
-module.exports = run;
+export default run;
