@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   SIGHUP: 1,
   SIGINT: 2,
   SIGQUIT: 3,
@@ -31,4 +31,4 @@ module.exports = {
   SIGPWR: 30,
   SIGSYS: 31,
   SIGRTMIN: 35,
-}
+};
