@@ -1,16 +1,18 @@
-var debug = require('debug')('nodemon');
-var path = require('path');
-var monitor = require('./monitor');
-var cli = require('./cli');
-var version = require('./version');
-var util = require('util');
-var utils = require('./utils');
-var bus = utils.bus;
-var help = require('./help');
+import Debug from 'debug';
+import path from 'path';
+import monitor from './monitor/index.js';
+import cli from './cli/index.js';
+import version from './version.js';
+import util from 'util';
+import utils from './utils/index.js';
+import help from './help/index.js';
 /** @type {import('..').NodemonEventConfig} */
-var config = require('./config');
-var spawn = require('./spawn');
-const defaults = require('./config/defaults')
+import config from './config/index.js';
+import spawn from './spawn.js';
+import defaults from './config/defaults.js';
+
+const debug = Debug('nodemon');
+const bus = utils.bus;
 var eventHandlers = {};
 
 // this is fairly dirty, but theoretically sound since it's part of the
@@ -57,12 +59,13 @@ function nodemon(settings) {
   }
 
   if (options.version) {
-    version().then(function (v) {
+    (async () => {
+      const v = await version();
       console.log(v);
       if (!config.required) {
         process.exit(0);
       }
-    });
+    })();
     return;
   }
 
@@ -313,5 +316,5 @@ bus.on('reset', function (done) {
 // expose the full config
 nodemon.config = config;
 
-module.exports = nodemon;
+export default nodemon;
 
