@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('../utils');
+import utils from '../utils/index.js';
 
 // internal
 var reEscComments = /\\#/g;
@@ -9,8 +9,6 @@ var reUnescapeComments = /\^\^/g;
 var reComments = /#.*$/;
 var reEscapeChars = /[.|\-[\]()\\]/g;
 var reAsterisk = /\*/g;
-
-module.exports = add;
 
 /**
  * Converts file patterns or regular expressions to nodemon
@@ -30,7 +28,7 @@ module.exports = add;
  * @param {String} which must be either "watch" or "ignore"
  * @param {String|RegExp} rule the actual rule.
  */
-function add(rules, which, rule) {
+export default function add(rules, which, rule) {
   if (!{ ignore: 1, watch: 1}[which]) {
     throw new Error('rules/index.js#add requires "ignore" or "watch" as the ' +
       'first argument');
@@ -46,7 +44,7 @@ function add(rules, which, rule) {
   // support the rule being a RegExp, but reformat it to
   // the custom :<regexp> format that we're working with.
   if (rule instanceof RegExp) {
-    // rule = ':' + rule.toString().replace(/^\/(.*?)\/$/g, '$1');
+    // rule = ':' + rule.toString().replace(/^\/(.*?)\/$\/, '$1');
     utils.log.error('RegExp format no longer supported, but globs are.');
     return;
   }
