@@ -1,7 +1,7 @@
 'use strict';
-var utils = require('../utils');
-var add = require('./add');
-var parse = require('./parse');
+import utils from '../utils/index.js';
+import add from './add.js';
+import parse from './parse.js';
 
 // exported
 var rules = { ignore: [], watch: [] };
@@ -33,7 +33,7 @@ function load(filename, callback) {
   });
 }
 
-module.exports = {
+export default {
   reset: function () { // just used for testing
     rules.ignore.length = rules.watch.length = 0;
     delete rules.ignore.re;
