@@ -1,5 +1,5 @@
 'use strict';
-var fs = require('fs');
+import fs from 'fs';
 
 /**
  * Parse the nodemon config file, supporting both old style
@@ -8,7 +8,7 @@ var fs = require('fs');
  * @param  {String}   filename
  * @param  {Function} callback
  */
-function parse(filename, callback) {
+export default function parse(filename, callback) {
   var rules = {
     ignore: [],
     watch: [],
@@ -38,6 +38,3 @@ function parse(filename, callback) {
     return callback(null, { raw: content.split(/\n/) });
   });
 }
-
-module.exports = parse;
-
