@@ -1,10 +1,11 @@
-const path = require('path');
-const utils = require('./utils');
+import path from 'path';
+import utils from './utils/index.js';
+import { spawn } from 'child_process';
+
 const merge = utils.merge;
 const bus = utils.bus;
-const spawn = require('child_process').spawn;
 
-module.exports = function spawnCommand(command, config, eventArgs) {
+export default function spawnCommand(command, config, eventArgs) {
   var stdio = ['pipe', 'pipe', 'pipe'];
 
   if (config.options.stdout) {
@@ -71,4 +72,4 @@ module.exports = function spawnCommand(command, config, eventArgs) {
       bus.stderr = child.stderr;
     }
   }
-};
+}
