@@ -1,12 +1,14 @@
-var events = require('events');
-var debug = require('debug')('nodemon');
-var util = require('util');
+import { EventEmitter } from 'events';
+import Debug from 'debug';
+import { inherits } from 'util';
+
+const debug = Debug('nodemon');
 
 var Bus = function () {
-  events.EventEmitter.call(this);
+  EventEmitter.call(this);
 };
 
-util.inherits(Bus, events.EventEmitter);
+inherits(Bus, EventEmitter);
 
 var bus = new Bus();
 
@@ -41,4 +43,4 @@ if (process.send) {
   };
 }
 
-module.exports = bus;
+export default bus;
