@@ -23,4 +23,4 @@ var re = new RegExp(('(' + reStr + ')').replace(/\[/g, '\\['), 'g');
 
 colour.strip = strip;
 
-module.exports = colour;
+export default colour;
