@@ -1,38 +1,21 @@
-var noop = function () { };
-var path = require('path');
-const semver = require('semver');
-var version = process.versions.node.split('.') || [null, null, null];
+import path from 'path';
+import { type } from 'os';
+import clone from './clone.js';
+import merge from './merge.js';
+import bus from './bus.js';
+import Logger from './log.js';
 
-var utils = (module.exports = {
-  semver: semver,
-  satisfies: test => semver.satisfies(process.versions.node, test),
-  version: {
-    major: parseInt(version[0] || 0, 10),
-    minor: parseInt(version[1] || 0, 10),
-    patch: parseInt(version[2] || 0, 10),
-  },
-  clone: require('./clone'),
-  merge: require('./merge'),
-  bus: require('./bus'),
+var noop = function () { };
+
+var utils = {
+  clone: clone,
+  merge: merge,
+  bus: bus,
   isWindows: process.platform === 'win32',
   isMac: process.platform === 'darwin',
   isLinux: process.platform === 'linux',
-  isIBMi: require('os').type() === 'OS400',
-  isRequired: (function () {
-    var p = module.parent;
-    while (p) {
-      // in electron.js engine it happens
-      if (!p.filename) {
-        return true;
-      }
-      if (p.filename.indexOf('bin' + path.sep + 'nodemon.js') !== -1) {
-        return false;
-      }
-      p = p.parent;
-    }
-
-    return true;
-  })(),
+  isIBMi: type() === 'OS400',
+  isRequired: false, // In ES modules, we assume it's always required
   home: process.env.HOME || process.env.HOMEPATH,
   quiet: function () {
     // nukes the logging
@@ -80,9 +63,9 @@ var utils = (module.exports = {
       .join(' ')
       .trim();
   },
-});
+};
 
-utils.log = require('./log')(utils.isRequired);
+utils.log = new Logger(utils.isRequired);
 
 Object.defineProperty(utils, 'debug', {
   set: function (value) {
@@ -101,3 +84,5 @@ Object.defineProperty(utils, 'colours', {
     return this.log.useColours;
   },
 });
+
+export default utils;
