@@ -1,5 +1,6 @@
-var colour = require('./colour');
-var bus = require('./bus');
+import colour from './colour.js';
+import bus from './bus.js';
+
 var required = false;
 var useColours = true;
 
@@ -79,4 +80,4 @@ Object.defineProperty(Logger.prototype, 'useColours', {
   },
 });
 
-module.exports = Logger;
+export default Logger;
