@@ -1,6 +1,4 @@
-var clone = require('./clone');
-
-module.exports = merge;
+import clone from './clone.js';
 
 function typesMatch(a, b) {
   return (typeof a === typeof b) && (Array.isArray(a) === Array.isArray(b));
@@ -12,7 +10,7 @@ function typesMatch(a, b) {
  * @param  {Object} target   [description]
  * @return {Object}          [description]
  */
-function merge(source, target, result) {
+export default function merge(source, target, result) {
   if (result === undefined) {
     result = clone(source);
   }
