@@ -1,5 +1,3 @@
-module.exports = command;
-
 /**
  * command constructs the executable command to run in a shell including the
  * user script, the command arguments.
@@ -15,7 +13,7 @@ module.exports = command;
  * @return {Object}          an object with the node executable and the
  *                           arguments to the command
  */
-function command(settings) {
+export default function command(settings) {
   var options = settings.execOptions;
   var executable = options.exec;
   var args = [];
