@@ -1,1 +1,1 @@
-module.exports = require('./nodemon');
+export { default } from './nodemon.js';
