@@ -1,15 +1,15 @@
-module.exports.watch = watch;
-module.exports.resetWatchers = resetWatchers;
+import Debug from 'debug';
+import chokidar from 'chokidar';
+import undefsafe from 'undefsafe';
+import path from 'path';
+import config from '../config/index.js';
+import utils from '../utils/index.js';
+import match, { rulesToMonitor } from './match.js';
 
-var debug = require('debug')('nodemon:watch');
-var debugRoot = require('debug')('nodemon');
-var chokidar = require('chokidar');
-var undefsafe = require('undefsafe');
-var config = require('../config');
-var path = require('path');
-var utils = require('../utils');
-var bus = utils.bus;
-var match = require('./match');
+const debug = Debug('nodemon:watch');
+const debugRoot = Debug('nodemon');
+const bus = utils.bus;
+
 var watchers = [];
 var debouncedBus;
 
@@ -23,7 +23,7 @@ function resetWatchers() {
   watchers = [];
 }
 
-function watch() {
+async function watch() {
   if (watchers.length) {
     debug('early exit on watch, still watching (%s)', watchers.length);
     return;
@@ -37,101 +37,101 @@ function watch() {
 
   var watchedFiles = [];
 
-  const promise = new Promise(function (resolve) {
-    const dotFilePattern = /[/\\]\./;
-    var ignored = match.rulesToMonitor(
-      [], // not needed
-      Array.from(rootIgnored),
-      config
-    ).map(pattern => pattern.slice(1));
+  try {
+    const total = await new Promise(function (resolve) {
+      const dotFilePattern = /[\/\\]\./;
+      var ignored = rulesToMonitor(
+        [], // not needed
+        Array.from(rootIgnored),
+        config
+      ).map(pattern => pattern.slice(1));
 
-    const addDotFile = dirs.filter(dir => dir.match(dotFilePattern));
+      const addDotFile = dirs.filter(dir => dir.match(dotFilePattern));
 
-    // don't ignore dotfiles if explicitly watched.
-    if (addDotFile.length === 0) {
-      ignored.push(dotFilePattern);
-    }
-
-    var watchOptions = {
-      ignorePermissionErrors: true,
-      ignored: ignored,
-      persistent: true,
-      usePolling: config.options.legacyWatch || false,
-      interval: config.options.pollingInterval,
-      // note to future developer: I've gone back and forth on adding `cwd`
-      // to the props and in some cases it fixes bugs but typically it causes
-      // bugs elsewhere (since nodemon is used is so many ways). the final
-      // decision is to *not* use it at all and work around it
-      // cwd: ...
-    };
-
-    if (utils.isWindows) {
-      watchOptions.disableGlobbing = true;
-    }
-
-    if (utils.isIBMi) {
-      watchOptions.usePolling = true;
-    }
-
-    if (process.env.TEST) {
-      watchOptions.useFsEvents = false;
-    }
-
-    var watcher = chokidar.watch(
-      dirs,
-      Object.assign({}, watchOptions, config.options.watchOptions || {})
-    );
-
-    watcher.ready = false;
-
-    var total = 0;
-
-    watcher.on('change', filterAndRestart);
-    watcher.on('unlink', filterAndRestart);
-    watcher.on('add', function (file) {
-      if (watcher.ready) {
-        return filterAndRestart(file);
+      // don't ignore dotfiles if explicitly watched.
+      if (addDotFile.length === 0) {
+        ignored.push(dotFilePattern);
       }
 
-      watchedFiles.push(file);
-      bus.emit('watching', file);
-      debug('chokidar watching: %s', file);
-    });
-    watcher.on('ready', function () {
-      watchedFiles = Array.from(new Set(watchedFiles)); // ensure no dupes
-      total = watchedFiles.length;
-      watcher.ready = true;
-      resolve(total);
-      debugRoot('watch is complete');
-    });
+      var watchOptions = {
+        ignorePermissionErrors: true,
+        ignored: ignored,
+        persistent: true,
+        usePolling: config.options.legacyWatch || false,
+        interval: config.options.pollingInterval,
+        // note to future developer: I've gone back and forth on adding `cwd`
+        // to the props and in some cases it fixes bugs but typically it causes
+        // bugs elsewhere (since nodemon is used is so many ways). the final
+        // decision is to *not* use it at all and work around it
+        // cwd: ...
+      };
 
-    watcher.on('error', function (error) {
-      if (error.code === 'EINVAL') {
-        utils.log.error(
-          'Internal watch failed. Likely cause: too many ' +
-          'files being watched (perhaps from the root of a drive?\n' +
-          'See https://github.com/paulmillr/chokidar/issues/229 for details'
-        );
-      } else {
-        utils.log.error('Internal watch failed: ' + error.message);
-        process.exit(1);
+      if (utils.isWindows) {
+        watchOptions.disableGlobbing = true;
       }
+
+      if (utils.isIBMi) {
+        watchOptions.usePolling = true;
+      }
+
+      if (process.env.TEST) {
+        watchOptions.useFsEvents = false;
+      }
+
+      var watcher = chokidar.watch(
+        dirs,
+        Object.assign({}, watchOptions, config.options.watchOptions || {})
+      );
+
+      watcher.ready = false;
+
+      var total = 0;
+
+      watcher.on('change', filterAndRestart);
+      watcher.on('unlink', filterAndRestart);
+      watcher.on('add', function (file) {
+        if (watcher.ready) {
+          return filterAndRestart(file);
+        }
+
+        watchedFiles.push(file);
+        bus.emit('watching', file);
+        debug('chokidar watching: %s', file);
+      });
+      watcher.on('ready', function () {
+        watchedFiles = Array.from(new Set(watchedFiles)); // ensure no dupes
+        total = watchedFiles.length;
+        watcher.ready = true;
+        resolve(total);
+        debugRoot('watch is complete');
+      });
+
+      watcher.on('error', function (error) {
+        if (error.code === 'EINVAL') {
+          utils.log.error(
+            'Internal watch failed. Likely cause: too many ' +
+            'files being watched (perhaps from the root of a drive?\n' +
+            'See https://github.com/paulmillr/chokidar/issues/229 for details'
+          );
+        } else {
+          utils.log.error('Internal watch failed: ' + error.message);
+          process.exit(1);
+        }
+      });
+
+      watchers.push(watcher);
     });
 
-    watchers.push(watcher);
-  });
-
-  return promise.catch(e => {
+    utils.log.detail(`watching ${watchedFiles.length} file${
+      watchedFiles.length === 1 ? '' : 's'}`);
+    return watchedFiles;
+  } catch (e) {
     // this is a core error and it should break nodemon - so I have to break
     // out of a promise using the setTimeout
     setTimeout(() => {
       throw e;
     });
-  }).then(function () {
-    utils.log.detail(`watching ${watchedFiles.length} file${
-      watchedFiles.length === 1 ? '' : 's'}`);
-    return watchedFiles;
-  });
+  }
 }
 
 function filterAndRestart(files) {
@@ -168,7 +168,6 @@ function filterAndRestart(files) {
         return f[0].toUpperCase() + f.slice(1);
       });
     }
-
 
     debug('filterAndRestart on', files);
 
@@ -239,6 +238,8 @@ function debounce(fn, delay) {
     const context = this;
     const args = arguments;
     clearTimeout(timer);
-    timer = setTimeout(() =>fn.apply(context, args), delay);
+    timer = setTimeout(() => fn.apply(context, args), delay);
   };
 }
+
+export { watch, resetWatchers };
