@@ -1,7 +1,5 @@
-module.exports = clone;
-
 // via http://stackoverflow.com/a/728694/22617
-function clone(obj) {
+export default function clone(obj) {
   // Handle the 3 simple types, and null or undefined
   if (null === obj || 'object' !== typeof obj) {
     return obj;
