@@ -1,89 +1,87 @@
-module.exports = version;
-module.exports.pin = pin;
+import fs from 'fs';
+import path from 'path';
+import { exec } from 'child_process';
+import { promisify } from 'util';
+import { fileURLToPath } from 'url';
+import { createRequire } from 'module';
 
-var fs = require('fs');
-var path = require('path');
-var exec = require('child_process').exec;
+const require = createRequire(import.meta.url);
+const execAsync = promisify(exec);
+const statAsync = promisify(fs.stat);
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
 var root = null;
 
-function pin() {
-  return version().then(function (v) {
-    version.pinned = v;
-  });
+async function pin() {
+  const v = await version();
+  version.pinned = v;
 }
 
-function version(callback) {
-  // first find the package.json as this will be our root
-  var promise = findPackage(path.dirname(module.parent.filename))
-    .then(function (dir) {
-      // now try to load the package
-      var v = require(path.resolve(dir, 'package.json')).version;
+async function version(callback) {
+  try {
+    // first find the package.json as this will be our root
+    const dir = await findPackage(__dirname);
+    
+    // now try to load the package
+    const v = require(path.resolve(dir, 'package.json')).version;
 
-      if (v && v !== '0.0.0-development') {
-        return v;
-      }
+    if (v && v !== '0.0.0-development') {
+      const result = v;
+      if (callback) callback(null, result);
+      return result;
+    }
 
-      root = dir;
+    root = dir;
 
-      // else we're in development, give the commit out
-      // get the last commit and whether the working dir is dirty
-      var promises = [
-        branch().catch(function () { return 'master'; }),
-        commit().catch(function () { return '<none>'; }),
-        dirty().catch(function () { return 0; }),
-      ];
+    // else we're in development, give the commit out
+    // get the last commit and whether the working dir is dirty
+    const [branchName, commitHash, dirtyCount] = await Promise.all([
+      branch().catch(() => 'master'),
+      commit().catch(() => '<none>'),
+      dirty().catch(() => 0),
+    ]);
 
-      // use the cached result as the export
-      return Promise.all(promises).then(function (res) {
-        var branch = res[0];
-        var commit = res[1];
-        var dirtyCount = parseInt(res[2], 10);
-        var curr = branch + ': ' + commit;
-        if (dirtyCount !== 0) {
-          curr += ' (' + dirtyCount + ' dirty files)';
-        }
+    const dirtyCountInt = parseInt(dirtyCount, 10);
+    let curr = branchName + ': ' + commitHash;
+    if (dirtyCountInt !== 0) {
+      curr += ' (' + dirtyCountInt + ' dirty files)';
+    }
 
-        return curr;
-      });
-    }).catch(function (error) {
-      console.log(error.stack);
-      throw error;
-    });
-
-  if (callback) {
-    promise.then(function (res) {
-      callback(null, res);
-    }, callback);
+    const result = curr;
+    if (callback) callback(null, result);
+    return result;
+  } catch (error) {
+    console.log(error.stack);
+    if (callback) callback(error);
+    throw error;
   }
-
-  return promise;
 }
 
-function findPackage(dir) {
+async function findPackage(dir) {
   if (dir === '/') {
-    return Promise.reject(new Error('package not found'));
+    throw new Error('package not found');
   }
-  return new Promise(function (resolve) {
-    fs.stat(path.resolve(dir, 'package.json'), function (error, exists) {
-      if (error || !exists) {
-        return resolve(findPackage(path.resolve(dir, '..')));
-      }
-
-      resolve(dir);
-    });
-  });
+  
+  try {
+    const exists = await statAsync(path.resolve(dir, 'package.json'));
+    if (exists) {
+      return dir;
+    }
+  } catch (error) {
+    // file doesn't exist, continue searching
+  }
+  
+  return await findPackage(path.resolve(dir, '..'));
 }
 
-function command(cmd) {
-  return new Promise(function (resolve, reject) {
-    exec(cmd, { cwd: root }, function (err, stdout, stderr) {
-      var error = stderr.trim();
-      if (error) {
-        return reject(new Error(error));
-      }
-      resolve(stdout.split('\n').join(''));
-    });
-  });
+async function command(cmd) {
+  const { stdout, stderr } = await execAsync(cmd, { cwd: root });
+  const error = stderr.trim();
+  if (error) {
+    throw new Error(error);
+  }
+  return stdout.split('\n').join('');
 }
 
 function commit() {
@@ -98,3 +96,8 @@ function dirty() {
   return command('expr $(git status --porcelain 2>/dev/null| ' +
     'egrep "^(M| M)" | wc -l)');
 }
+
+version.pin = pin;
+
+export default version;
+export { pin };
